Update caregiver_locator_form for exception handling
Added exception handling for retrieving "CaregiverLocator" instances in caregiver_locator_form.py. Previously, if no instance was found in the database with the given subject identifier, the code would raise a server error. The logic has been changed so that now, if no instances are found, the function will silently pass. This ensures there will not be a server error while looking for non-existing instances and improves system reliability.