:ambulance: feature(Forms): Refactor subject identifier check logic in forms
Adjusted the condition that checks for the existence of a 'subject_identifier' in caregiver form. Instead of getting the identifier from 'initial', it is now extracted from 'cleaned_data'. This change prevents a potential error that could occur when the initial data does not contain the 'subject_identifier'. The validation logic has been wrapped inside an if block that checks for the presence of 'subject_identifier', thereby avoiding needless processing when 'subject_identifier' is not provided. This change aims at enhancing application's performance and reliability.