:recycle: Refactor(hiv_disclosure_status_admin): Handle visit_model.DoesNotExist exception
The code searching for a visit_model object in the hiv_disclosure_status_admin.py file has been refactored to include a try-except clause. This is to handle instances where the desired visit_model object does not exist, hence improving the robustness of the overall system. Moreover, it provides a more detailed and informative response when such instance occurs.