Remove unique constraint for caregiver identifier
The unique constraint for the 'Study Caregiver Subject Identifier' in the childbirth form was removed. This was necessary to be able to link multiple caregivers with the same identifier to different maternal identifiers, allowing for better data coverage and flexibility in handling different cases. Potential data integrity issues have been anticipated and will need to be considered during testing.