:ambulance: fix(consent form): Ignore non-existent screening on consent form
Removed the ValueError that arose when trying to fetch a non-existing screening object from the database in the Subject Consent Form. Previously, if a screening identifier was not found in the database, the system would raise a ValueError and halt execution. This commit allows the process to continue even if the specific screening does not exist - thus, not disrupting the flow of the application.