Fix child consent version reference in caregiver_child_consent
Changed the variable that ‘child_consent_version’ data is taken from, in caregiver_child_consent model. Previously, it was referencing itself causing an exception when 'is_eligible' was True and 'subject_identifier' and 'version' was not present. This fix calls the correct 'caregiver_config' for fetching 'child_consent_version'. This will ensure the correct consent version number is set under the specified conditions.