Update 'score' field in 'HITSScreening'
The 'score' field in 'HITSScreening' model, form, and admin views has been updated. An additional attribute 'score' is now displayed in the admin view. In addition, the label for this field in the form has been changed and is no longer required. The verbose name of 'score' in the model is now 'Total HITS Score.'