Update Python versions and reorganize workflow in Django.yml
Updated the versions of Python used in the .github/workflows/django.yml file from 3.9, 3.10, and 3.11 to 3.7, 3.8, and 3.9. Additionally, simplified and reorganized the steps for installing coverage. The update not only ensures compatibility with different Python versions but also increases efficiency and clarity of the actions carried out in each step.