Revise visit sequence logic in Flourish Caregiver project
The visit sequence calculations for the Flourish Caregiver project have been substantially revised and expanded. This includes the establishment of visit sequence errors, exclusion of visit codes, and implementing various methods for obtaining previous visits. This ensures more efficient tracking of participant's visits and aids in maintaining a correct sequence of visits.