(Refactor): Refactor subject consent admin and create option generator
This commit includes modifications for subject_consent_admin.py in an effort to optimize code, moving the generation of initial values to a dedicated function. It also includes the addition of methods for fetching information from locator and pre_flourish_consent model objects, as well as a separate function for generating participant options. The set_initials function has been moved into a utility module for broader use.

Signed-off-by: nmunatsibw <mcturnergarai@gmail.com>