(Update): Adjust the list of test options in Flourish Caregiver and Child apps
A 'None' option was added to the list of test types in the Flourish Caregiver app. Conversely, this option was removed from the list of test types in the Flourish Child app. These changes are meant to ensure the lists of tests available in each app reflect the specific needs and contexts of the caregivers and children respectively.

Signed-off-by: nmunatsibw <mcturnergarai@gmail.com>