(refactor): Update model and constants import order
This commit cleans up the import order in 'cohort.py' for better readability and also removes some extraneous whitespaces. A list of ignored schedule names is added to more specifically target the filtering process in 'check_current_cohort' function. The filtered result is now determined by checking if 'schedule_name' is in the ignored list, enhancing code clarity and maintainability.

Signed-off-by: nmunatsibw <mcturnergarai@gmail.com>