(refactor): Update list and model definitions
The code has been updated to improve clarity and integrity. In list_data.py, the 'None' constant is now correctly represented as NONE. In choices.py, a new constant tuple YES_NO_AR_OTHER has been added. Lastly, in flourish_tb_screening_mixin.py, the prompt for 'household_diagnosed_with_tb' has been changed to mention the diagnosis of TB in the household, instead of the evaluation of TB in a clinic.

Signed-off-by: nmunatsibw <mcturnergarai@gmail.com>