(refactor): Simplify caregiver TB screening and diagnosis functionalities
Changes in this commit include the simplification of the evaluation for TB in the TBDiagnosis class. This includes the structuring of the evaluate_for_tb method and the creation of the evaluate_screening_conditions method which serves to tidy up the logic of evaluating whether the subject meets the conditions for TB screening.

Additionally, the CaregiverTBScreeningForm was amended to include a method named get_keys_before which retrieves all dictionary keys before a specified key. This change improves the structure and readability of the form class. Import statements from various other classes were also adjusted.

Signed-off-by: nmunatsibw <mcturnergarai@gmail.com>