(tests): Simplify childhood lead exposure risk test
The complex "make_recipe" method to create a maternalvisit object was removed from the childhood lead exposure risk test. This was unnecessary and overly complicated. Now, only the CrfMetadata object check remains. The visit code was also corrected from '3005M' to '3004M' in this check. The test is much simplified and accurately checks the '3004M' visit entry status.

Signed-off-by: nmunatsibw <mcturnergarai@gmail.com>