(feature-update): Remove null from model and update TB evaluation conditions
This commit removes the null attribute from the 'evaluated_for_tb' field in the 'flourish_tb_screening_mixin' model and sets the default to an empty string. It also updates the Tuberculosis (TB) evaluation conditions in 'tb_diagnosis.py' to include a check for diagnosis in the household. The update aims at increasing accuracy and filling data gaps during TB screening and evaluation.

Signed-off-by: nmunatsibw <mcturnergarai@gmail.com>